Improved history handling after connection was broken, history is no longer loaded again
git-svn-id: file:///home/svn/repos/tigase-monitor/trunk@67 f3d9453a-065d-4d14-b440-5a2b23bfa02b

+++ bin/prepare-dist.sh
@@ -1,7 +1,12 @@
 #!/bin/bash
 
+if [ "$1" == "" ] ; then
+  echo "Missing parameter - client id."
+  exit 2
+fi
+
 DATE=`date +"%Y%m%d"`
-DIR="tigase-monitor_$DATE"
+DIR="tigase-monitor_$1_$DATE"
 
 rm -rf $DIR $DIR.tgz
 
@@ -10,12 +15,14 @@ mkdir $DIR/etc
 mkdir $DIR/libs
 mkdir $DIR/sounds
 
-cp target/tigase-monitor*.jar $DIR/tigase-monitor.jar
-jar ufm $DIR/tigase-monitor.jar bin/manifest.txt
-cp ../server/jars/tigase-server.jar $DIR/libs/
-cp lib/*.jar $DIR/libs/
+#set -x
+
+cp target/tigase-monitor*jar-with-dependencies.jar $DIR/libs/tigase-monitor.jar
 cp sounds/* $DIR/sounds/
 cp bin/monitor.sh $DIR/
-cp bin/monitor.properties $DIR/etc/
+cp etc/${1}-init.properties $DIR/etc/monitor.properties
+cp ../client-licenses/${1}-monitor.licence $DIR/etc/monitor.licence
 
 tar -czf $DIR.tgz $DIR
+
+mv -f $DIR.tgz packs/
